+++ src/lib.rs
@@ -6,3 +6,4 @@ pub mod perf_events;
 pub mod profiler;
 pub mod unwind_info;
 pub mod usym;
+pub mod profile;
